Extract props extraction from container
