+++ src/app/legacy/containers/MediaPlayer/helpers/propsInference/index.js
@@ -1,0 +1,12 @@
+import getDefaultProps from './default';
+import livePageProps from './livePage';
+
+const getPlayerProps = ({ pageType, ...params }) => {
+  const propsHandler = {
+    live: livePageProps,
+  }[pageType];
+
+  return propsHandler ? propsHandler(params) : getDefaultProps(params);
+};
+
+export default getPlayerProps;
