+++ src/app/legacy/containers/MediaPlayer/helpers/propsInference/livePage/index.js
@@ -1,0 +1,3 @@
+export default () => {
+  return { aresMediaBlock: null };
+};
