+++ src/app/legacy/containers/MediaPlayer/index.jsx
@@ -1,8 +1,5 @@
 import React, { useContext } from 'react';
 import { string, bool } from 'prop-types';
-import moment from 'moment-timezone';
-import pathOr from 'ramda/src/pathOr';
-import path from 'ramda/src/path';
 import Figure from '#psammead/psammead-figure/src';
 import useLocation from '#hooks/useLocation';
 import {
@@ -20,10 +17,6 @@ import {
 import getEmbedUrl, {
   makeAbsolute,
 } from '#lib/utilities/getUrlHelpers/getEmbedUrl';
-import { getPlaceholderSrcSet } from '#lib/utilities/srcSet';
-import filterForBlockType from '#lib/utilities/blockHandlers';
-import formatDuration from '#lib/utilities/formatDuration';
-import buildIChefURL from '#lib/utilities/ichefURL';
 import { RequestContext } from '#contexts/RequestContext';
 import {
   mediaPlayerPropTypes,
@@ -33,8 +26,8 @@ import { ServiceContext } from '../../../contexts/ServiceContext';
 import Metadata from './Metadata';
 import Caption from '../Caption';
 import logMissingMediaId from './helpers/logMissingMediaId';
+import getPlayerProps from './helpers/propsInference';
 
-const DEFAULT_WIDTH = 512;
 const MediaPlayerContainer = ({
   blocks,
   assetId,
@@ -45,79 +38,39 @@ const MediaPlayerContainer = ({
   showLoadingImage,
   showCaption,
 }) => {
-  const { isAmp } = useContext(RequestContext);
+  const { isAmp, pageType } = useContext(RequestContext);
   const { lang, translations, service } = useContext(ServiceContext);
 
   if (!blocks) {
     return null;
   }
 
-  const aresMediaBlock = filterForBlockType(blocks, 'aresMedia');
-  const articleCaptionBlock = filterForBlockType(blocks, 'caption');
-  const cpsCaptionBlock = filterForBlockType(
-    path(['model', 'blocks'], aresMediaBlock),
-    'caption',
-  );
-  const captionBlock = articleCaptionBlock || cpsCaptionBlock;
+  const {
+    aresMediaBlock,
+    blockId,
+    captionBlock,
+    embedUrlParams,
+    iframeTitle,
+    mediaInfo,
+    placeholderSrc,
+    placeholderSrcset,
+    translatedExpiredContentMessage,
+    translatedNoJSMessage,
+    versionId,
+  } = getPlayerProps({
+    assetId,
+    assetType,
+    blocks,
+    isAmp,
+    isLegacyMedia,
+    lang,
+    pageType,
+    translations,
+  });
 
   if (!aresMediaBlock) {
     return null;
   }
-
-  const { originCode, locator } = pathOr(
-    {},
-    ['model', 'blocks', 1, 'model', 'blocks', 0, 'model'],
-    aresMediaBlock,
-  );
-  const versionId = path(
-    ['model', 'blocks', 0, 'model', 'versions', 0, 'versionId'],
-    aresMediaBlock,
-  );
-  const blockId = path(
-    ['model', 'blocks', 0, 'model', 'blockId'],
-    aresMediaBlock,
-  );
-
-  const format = path(
-    ['model', 'blocks', 0, 'model', 'format'],
-    aresMediaBlock,
-  );
-  const rawDuration = path(
-    ['model', 'blocks', 0, 'model', 'versions', 0, 'duration'],
-    aresMediaBlock,
-  );
-  const duration = moment.duration(rawDuration, 'seconds');
-  const durationSpokenPrefix = pathOr(
-    'Duration',
-    ['media', 'duration'],
-    translations,
-  );
-  const separator = ',';
-
-  const mediaInfo = {
-    title: path(['model', 'blocks', 0, 'model', 'title'], aresMediaBlock),
-    duration: formatDuration({ duration, padMinutes: true }),
-    durationSpoken: `${durationSpokenPrefix} ${formatDuration({
-      duration,
-      separator,
-    })}`,
-    datetime: path(
-      ['model', 'blocks', 0, 'model', 'versions', 0, 'durationISO8601'],
-      aresMediaBlock,
-    ),
-    type: format === 'audio' ? 'audio' : 'video',
-    guidanceMessage: path(
-      ['model', 'blocks', 0, 'model', 'versions', 0, 'warnings', 'short'],
-      aresMediaBlock,
-    ),
-  };
-
-  const placeholderSrcset = getPlaceholderSrcSet({ originCode, locator });
-  const placeholderSrc = buildIChefURL({
-    originCode,
-    locator,
-    resolution: DEFAULT_WIDTH,
-  });
 
   const landscapeRatio = '56.25%'; // (9/16)*100 = 16:9
   const StyledMessageContainer = styled.div`
@@ -135,16 +88,6 @@ const MediaPlayerContainer = ({
     padding-bottom: ${GEL_SPACING_TRPL};
     width: 100%;
   `;
-
-  const noJsMessage = `This ${mediaInfo.type} cannot play in your browser. Please enable JavaScript or try a different browser.`;
-  const contentNotAvailableMessage = `This content is no longer available`;
-
-  const translatedNoJSMessage =
-    path(['media', 'noJs'], translations) || noJsMessage;
-
-  const translatedExpiredContentMessage =
-    path(['media', 'contentExpired'], translations) ||
-    contentNotAvailableMessage;
 
   const mediaIsValid = available && (versionId || blockId);
   if (!mediaIsValid) {
@@ -169,17 +112,9 @@ const MediaPlayerContainer = ({
   const location = useLocation();
 
   const embedSource = getEmbedUrl({
-    mediaId: `${assetId}/${isLegacyMedia ? blockId : versionId}/${lang}`,
-    type: assetType,
-    isAmp,
+    ...embedUrlParams,
     queryString: location.search,
   });
-
-  const iframeTitle = pathOr(
-    'Media player',
-    ['mediaAssetPage', 'mediaPlayer'],
-    translations,
-  );
 
   const caption = captionBlock ? (
     <Caption block={captionBlock} type={mediaInfo.type} service={service} />
