+++ ws-nextjs-app/pages/[service]/live/[id]/Post/index.tsx
@@ -9,6 +9,7 @@ import Text from '#app/components/Text';
 import Blocks from '#app/legacy/containers/Blocks';
 import Paragraph from '#app/legacy/containers/Paragraph';
 import UnorderedList from '#app/legacy/containers/BulletedList';
+import ArticleMediaPlayer from '#app/legacy/containers/ArticleMediaPlayer';
 import VisuallyHiddenText from '#app/components/VisuallyHiddenText';
 import {
   Post as PostType,
@@ -130,6 +131,11 @@ const PostContent = ({ contentBlocks }: { contentBlocks: OptimoBlock[] }) => {
     ),
     image: (props: { blocks: OptimoBlock[] }) => (
       <ImageWithCaption {...props} sizes="(min-width: 1008px) 760px, 100vw" />
+    ),
+    video: (props: ComponentToRenderProps) => (
+      <div>
+        <ArticleMediaPlayer blocks={props.blocks} />
+      </div>
     ),
     social: SocialEmbedContainer,
   };
